mainDocker/api/requirements.txt
Django==3.0.4
django-rest-knox==4.1.0
djangorestframework==3.11.0
